refactor(controller/wavs): extract waveform loading orchestration
